Implemented requested changes, filePath can is passed in during object construction. More detailed error handling. Function and package refactoring. Removed User toString method
