+++ src/main/java/entity/User.java
@@ -39,26 +39,4 @@ public class User {
     public void setPassword(String password) {
         this.password = password;
     }
-
-
-    @Override
-    public String toString() {
-        String listNames = eventLists.stream()
-                .filter(list -> !list.getName().equals("master_list"))
-                .map(list -> String.format("%s (%d events)", list.getName(), list.getEvents().size()))
-                .collect(Collectors.joining("; "));
-
-        if (listNames.isEmpty()) {
-            listNames = "None";
-        }
-
-        // Count events in the master list
-        int masterListCount = masterList != null ? masterList.getEvents().size() : 0;
-
-        return "--- USER ACCOUNT DETAILS ---\n" +
-                "Username:      " + username + "\n" +
-                "Master List:   " + masterList.getName() + " (" + masterListCount + " events)\n" +
-                "Other Lists:   " + listNames + "\n" +
-                "----------------------------";
-    }
 }
